as + ld: hande la

+++ include/elf/symbols.h
@@ -33,7 +33,7 @@ typedef struct
     unsigned char st_info;
     unsigned char st_other;
     uint16_t  st_shndx;
-    uint16_t st_value;
+    uint64_t st_value;
     uint64_t st_size;
 } Elf64_Sym;
